backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dan Tcheche <dtcheche@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic-settings = "^2.7.0"
ruff = "^0.8.4"

[tool.ruff]
line-length = 79
exclude = [
    "__init__.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
